OV-52: * rename
