+++ backend/src/bundles/chat/chat.ts
@@ -1,8 +1,8 @@
 import { logger } from '~/common/logger/logger.js';
 import { openAIService } from '~/common/services/services.js';
 
-import { ChatService } from './char.service.js';
 import { ChatController } from './chat.controller.js';
+import { ChatService } from './chat.service.js';
 
 const chatService = new ChatService();
 const chatController = new ChatController(logger, openAIService, chatService);
